+++ SerenityLib/Engine/src/core/engine/renderer/Renderer.h
@@ -47,8 +47,8 @@ namespace priv{
             void _init();
             void _resize(uint width, uint height);
 
-            void _render(const double& dt, Viewport&, const bool mainRenderFunc = true, const GLuint display_fbo = 0, const GLuint display_rbo = 0);
-            void _render2DApi(const double& dt, Viewport&, const bool mainRenderFunc = true, const GLuint display_fbo = 0, const GLuint display_rbo = 0);
+            void _render(const float& dt, Viewport&, const bool mainRenderFunc = true, const GLuint display_fbo = 0, const GLuint display_rbo = 0);
+            void _render2DApi(const float& dt, Viewport&, const bool mainRenderFunc = true, const GLuint display_fbo = 0, const GLuint display_rbo = 0);
 
             void _onFullscreen(const unsigned int& width, const unsigned int& height);
             void _onOpenGLContextCreation(uint width,uint height,uint glslVersion,uint openglVersion);
