+++ SerenityLib/Engine/src/core/engine/renderer/particles/ParticleSystem.h
@@ -31,8 +31,8 @@ namespace Engine {
                 std::stack<size_t>                     m_ParticleFreelist;
                 std::mutex                             m_Mutex;
 
-                void internal_update_emitters(const double& dt);
-                void internal_update_particles(const double& dt);
+                void internal_update_emitters(const float& dt);
+                void internal_update_particles(const float& dt);
             public:
                 ParticleSystem();
                 ~ParticleSystem();
@@ -52,7 +52,7 @@ namespace Engine {
                 std::stack<size_t>&           getParticleEmittersFreelist();
                 std::stack<size_t>&           getParticlesFreelist();
 
-                void update(const double& dt);
+                void update(const float& dt);
                 void render(Camera& camera, ShaderProgram& program, GBuffer& gBuffer);
             };
     };
