+++ SerenityLib/Engine/src/core/engine/renderer/pipelines/DeferredPipeline.h
@@ -80,7 +80,7 @@ namespace Engine {
                 void render2DTexture(const Texture* texture, const glm::vec2& position, const glm::vec4& color, const float angle, const glm::vec2& scale, const float depth, const Alignment::Type& align, const glm::vec4& scissor = glm::vec4(-1.0f));
                 void render2DTriangle(const glm::vec2& pos, const glm::vec4& color, const float angle, const float width, const float height, const float depth, const Alignment::Type& align, const glm::vec4& scissor = glm::vec4(-1.0f));
 
-                void update(const double& dt);
+                void update(const float& dt);
                 void render();
         };
     };
