+++ SerenityLib/Engine/src/core/engine/renderer/pipelines/IRenderingPipeline.h
@@ -47,7 +47,7 @@ namespace Engine {
                 virtual void render2DTexture(const Texture* texture, const glm::vec2& position, const glm::vec4& color, const float angle, const glm::vec2& scale, const float depth, const Alignment::Type& align, const glm::vec4& scissor = glm::vec4(-1.0f)) = 0;
                 virtual void render2DTriangle(const glm::vec2& pos, const glm::vec4& color, const float angle, const float width, const float height, const float depth, const Alignment::Type& align, const glm::vec4& scissor = glm::vec4(-1.0f)) = 0;
 
-                virtual void update(const double& dt) = 0;
+                virtual void update(const float& dt) = 0;
                 virtual void render() = 0;
